refactor: centralize helper type declarations
