+++ src/features/auth/verification/type.ts
@@ -1,1 +1,1 @@
-export type { VerifyValues } from "../../../utils/authHelper/types";
+export type { VerifyValues } from "../../../lib/types";
