+++ src/lib/types/_apiHelperType.ts
@@ -1,5 +1,16 @@
 import type { Method } from "axios";
 
+export type ApiError = {
+  data?: unknown;
+  message: string;
+  status?: number;
+};
+
+export type ApiSuccess<T> = { data: T; headers: unknown; status: number };
+export type ApiResult<T> =
+  | { error: ApiError; ok: false }
+  | { ok: true; response: ApiSuccess<T> };
+
 export type RequestOptions<TRequest> = {
   accessToken?: string;
   apiPath: string;
