+++ src/lib/types/index.ts
@@ -1,2 +1,4 @@
+export type * from "./_apiHelperType";
+export type * from "./_authHelperType";
 export type * from "./_componentsType";
 // export type * from "./_pagesType";
